Adding filtering function for new all metadata endpoint
Summary:
Adds a filtering function for Metadata that allows querying over all
metadata for an application, using the new `queryable_value` field.

Test Plan: Tested in redwood via new API endpoint.

Reviewers: spang

Reviewed By: spang

Subscribers: khamidou, jenkins, juan

Differential Revision: https://phab.nylas.com/D3102

+++ inbox/models/metadata.py
@@ -47,7 +47,7 @@ class Metadata(MailSyncBase, HasPublicID, HasRevisions, UpdatedAtMixin,
 
     value = Column(JSON)
 
-    queryable_value = Column(Integer, nullable=True, index=True)
+    queryable_value = Column(BigInteger, nullable=True, index=True)
 
     version = Column(Integer, nullable=True, server_default='0')
